+++ lib/ex_arrow/native.ex
@@ -130,6 +130,8 @@ defmodule ExArrow.Native do
   def record_batch_from_column_binaries(_names, _binaries, _dtypes, _length),
     do: :erlang.nif_error(:nif_not_loaded)
 
+  def record_batch_concat(_batches), do: :erlang.nif_error(:nif_not_loaded)
+
   # Compute kernels
   def compute_filter(_batch_ref, _predicate_ref), do: :erlang.nif_error(:nif_not_loaded)
   def compute_project(_batch_ref, _column_names), do: :erlang.nif_error(:nif_not_loaded)